+++ app/frontend/src/loginContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+export const LoginContext = createContext({
+    loggedIn: false,
+    setLoggedIn: (_: boolean) => {}
+});
